[SIFISC-31071] Passé à la version 2 de ehcache.

+++ core/src/main/java/ch/vd/unireg/cache/CacheStats.java
@@ -35,7 +35,7 @@ public interface CacheStats {
 	/**
 	 * @return nombre maximum d'éléments en mémoire; ou <b>null</b> si cette information n'est pas applicable
 	 */
-	Integer getMaxElements();
+	Long getMaxElements();
 
 	String toString();
 }
