+++ core/src/main/java/ch/vd/unireg/cache/SimpleCacheStats.java
@@ -57,7 +57,7 @@ public class SimpleCacheStats implements CacheStats {
 	}
 
 	@Override
-	public Integer getMaxElements() {
+	public Long getMaxElements() {
 		return null;
 	}
